feat: get tree by id

+++ internal/service/domain/tree/tree.go
@@ -31,8 +31,8 @@ func (s *TreeService) GetAll(ctx context.Context) ([]*entities.Tree, error) {
   return trees, nil
 }
 
-func (s *TreeService) GetByID(ctx context.Context, id int32) (*entities.Tree, error) {
-  tree, err := s.treeRepo.GetByID(ctx, id)
+func (s *TreeService) GetByID(ctx context.Context, id int) (*entities.Tree, error) {
+  tree, err := s.treeRepo.GetByID(ctx, int32(id))
   if err != nil {
     return nil, handleError(err)
   }
